# Use checksums instead of fsyncs to avoid slow discovery caching on MacOS