… vendor/github.com/gregjones/httpcache/diskcache/diskcache.go
… vendor/modules.txt