@lavalamp 2022-07-06
Feedback:
1. Hash the whole file
2. Use a cryptographic hash (note that, if you don't like how big they are, it's OK to only use some of the bytes for the key)