@apelisse 2022-07-25
> So by changing this hash algorithm, all currently disk-cached files will no longer be found.

Yeah, we never use that for discovery (yet), and OpenAPI changes all the time. This is as bad as installing a CRD on a cluster. I don't know if it warrants a release note.