@negz 2022-07-27
@lavalamp @seans3 PTAL - we're now SHA256 summing all the things. 😄