@csantanapr 2022-08-18
@negz Could you provide a one-liner for release notes?