@negz 2022-08-18
@csantanapr Not sure how much I'd consider this a user facing change, but I've added a one-liner nonetheless. Thanks!